Move `SharedPacakge` from `ckb-launcher` to `ckb-shared`
