+++ shared/Cargo.toml
@@ -26,6 +26,7 @@ ckb-async-runtime = { path = "../util/runtime", version = "= 0.113.0-pre" }
 ckb-stop-handler = { path = "../util/stop-handler", version = "= 0.113.0-pre" }
 ckb-constant = { path = "../util/constant", version = "= 0.113.0-pre" }
 ckb-systemtime = { path = "../util/systemtime", version = "= 0.113.0-pre" }
+ckb-channel = { path = "../util/channel", version = "= 0.113.0-pre" }
 
 [dev-dependencies]
 ckb-systemtime = { path = "../util/systemtime", version = "= 0.113.0-pre", features = ["enable_faketime"] }
